requirements.txt
pytest==8.3.3
anthropic==0.34.2
astor==0.8.1
PyYAML==6.0.2
rich==13.8.1
pydantic==2.9.2
colorama==0.4.6
python-dotenv==1.0.1
tiktoken==0.7.0
pydantic==2.9.2
PyYAML==6.0.2
